[cluster-auth] Move server-side RPC auth framework into rigging
Make iris a consumer of a generic "protect a Connect/RPC app" framework
rather than its owner. The token verifiers, the per-request identity
plumbing, the authenticator stack (presented token > IAP assertion >
trusted loopback), and the Connect interceptors move from iris.rpc.auth
to a new rigging.server_auth — the server-side companion to
rigging.auth's client-side token providers. iris.rpc.auth keeps only what
is iris-specific: its RBAC policy (AuthzAction / POLICY / the dashboard
read allowlist / authorize*) and the session-cookie name.

Three seams are parameterized so the framework carries no service
semantics:
- IapAssertionVerifier takes a required role_resolver (no built-in
  dashboard-role default); iris supplies its own.
- extract_bearer_token / AuthInterceptor / NullAuthInterceptor take an
  opt-in cookie_name; iris passes SESSION_COOKIE where the dashboard
  honours a browser session.
- ControllerAuthPolicy -> RequestAuthPolicy (no longer controller-bound).

rigging stays a leaf (imports no higher layer); its server framework now
carries its own tests in lib/rigging/tests/test_server_auth.py, and iris
keeps only the RBAC tests.
