+++ lib/iris/tests/cluster/controller/test_container_profile.py
@@ -25,7 +25,7 @@ from iris.cluster.controller.service import ControllerServiceImpl
 from iris.cluster.controller.worker_health import WorkerHealthTracker
 from iris.cluster.types import JobName
 from iris.rpc import controller_pb2, job_pb2
-from iris.rpc.auth import VerifiedIdentity, _verified_identity
+from rigging.server_auth import VerifiedIdentity, _verified_identity
 
 from tests.cluster.controller.conftest import (
     MockController,
