+++ lib/iris/tests/test_budget.py
@@ -25,8 +25,8 @@ from iris.cluster.controller.service import ControllerServiceImpl
 from iris.cluster.controller.worker_health import WorkerHealthTracker
 from iris.cluster.types import JobName, UserBudgetDefaults, WorkerId
 from iris.rpc import controller_pb2, job_pb2
-from iris.rpc.auth import VerifiedIdentity, _verified_identity
 from iris.rpc.proto_display import PRIORITY_BAND_VALUES, priority_band_name, priority_band_value
+from rigging.server_auth import VerifiedIdentity, _verified_identity
 from rigging.timing import Timestamp
 
 from tests.cluster.controller.conftest import (
