+++ lib/rigging/pyproject.toml
@@ -13,6 +13,7 @@ dependencies = [
     "gcsfs>=2024.0.0",
     "google-auth>=2.0.0",
     "pyyaml>=6.0",
+    "requests>=2.0.0",
     "s3fs>=2024.0.0",
 ]
